backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
passlib[bcrypt]
python-multipart
python-jose
duckdb
redis
pyarrow
pandas
email-validator
argon2-cffi
google-generativeai
